Refactor existing mgmt. command to add all users to the Public Group Django's loaddata mgmt. command does not trigger the User model's save() method so no post_save signals are triggered
